requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pillow>=9.0.0
opencv-python>=4.7.0
librosa>=0.10.0
pytest>=7.3.0
tqdm>=4.65.0